requirements.txt
scikit-learn==0.24.1
pandas==1.2.4
adversarial-robustness-toolbox==1.8.0
numpy==1.19.5
scikit-lego==0.6.8
aix360==0.1.0
uq360==0.1.1
anytree==2.8.0
requests-oauthlib==1.3.0
pyyaml==5.4.1
dill==0.3.4
lime==0.2.0.1
requests==2.25.1
tensorflow==2.6.0
scipy==1.6.2